+++ cmd/savk/main.go
@@ -21,7 +21,7 @@ import (
 )
 
 var (
-	version                  = "0.1.3"
+	version                  = "0.1.4"
 	commit                   = "unknown"
 	buildDate                = "unknown"
 	newPathChecker           = func() collectors.PathChecker { return collectors.OSPathChecker{} }